requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
bcrypt==4.0.1
    # via flask-bcrypt
blinker==1.6.3
    # via flask
build==1.0.3
    # via pip-tools
click==8.1.7
    # via
    #   flask
    #   pip-tools
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-cors
    #   flask-sqlalchemy
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-cors==4.0.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyjwt==2.8.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
sqlalchemy==2.0.22
    # via flask-sqlalchemy
typing-extensions==4.8.0
    # via sqlalchemy
werkzeug==3.0.0
    # via flask
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
